+++ src/types/input.ts
@@ -41,4 +41,5 @@ export type InputProps = {
 	value?: string | number // additional
 	maxLength?: number // additional
 	onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void // additional
+	autoFocus?: boolean // additional
 }